+++ src/main/resources/assets/gtadditions/lang/en_us.lang
@@ -2692,6 +2692,10 @@ gtadditions.multiblock.universal.tooltip.2=EUt Multiplier: §e%.1f§r
 gtadditions.multiblock.universal.tooltip.3=Duration Multiplier: §e%.1f§r
 gtadditions.multiblock.universal.tooltip.4=Max Parallel: §e%d§7 Per Overclocking Tier
 gtadditions.multiblock.universal.tooltip.5=Boost Chance: §e%d%%§r
+gtadditions.multiblock.universal.distinct=Distinct Buses:
+gtadditions.multiblock.universal.distinct.yes=§aYes
+gtadditions.multiblock.universal.distinct.no=§cNo
+gtadditions.multiblock.universal.distinct.info=If enabled, each bus will be treated as fully distinct from eachother for recipe lookup. Useful for example for Extruder Shapes, Laser Lenses, etc..
 gtadditions.multiblock.fusion_reactor.heat=Heat: %d
 gtadditions.multiblock.fusion_reactor.tooltip.1=EU To Start: %s
 gtadditions.multiblock.central_monitor.height=Screen Height: %d
